backend/login/requirements.txt
Flask
pymongo
werkzeug
pyjwt
flask_cors
pytest
requests
Flask-Mail
python-dotenv
PyJWT